requirements.txt
streamlit
streamlit-option-menu 
pymongo
dnspython
python-dotenv
google-generativeai